restructure utils; guidance page branchOfService
